# api/openapi/v1.yaml
openapi: 3.0.3
info:
  title: Tagona API
  version: 1.8.0
  description: |
    Storage for collections of objects with sparse boolean tags that are evaluated on demand.

    **Contract version.** This document describes the `/v1` API. `info.version` is the
    semantic version of this contract: additions bump the minor version. While Tagona is a draft, a
    breaking change is announced in the changelog and also bumps the minor version; once it is stable,
    breaking changes will ship as `/v2` with its own document.

    **Authentication.** Every `/v1/collections...` and `/v1/taggers` request needs `Authorization: Bearer <api key>`.
    Keys are minted with the admin endpoints, which use HTTP Basic auth.
    The documentation endpoints (`/v1/docs`, `/v1/openapi.json`, `/v1/openapi.yaml`) are public.

    **Limits.** Request bodies are capped (`413 payload_too_large`).

    **Retention.** Objects are stored permanently unless they have an expiry: either the server's
    default TTL (off by default) or the `ttl_seconds` upload parameter. Expired objects disappear
    from the API immediately and are removed from storage by a background sweeper. The exact rules
    are described under *Upload an object*.

    **Tags are sparse.** An object only has a value for a tag after the tag was evaluated for it
    (by a query or a tags request). A tag that was not evaluated is *unknown*. Both tag-aware
...
    tagging engine.

...

...

...
    known) are not affected.

...

...
  license:
...
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
  /v1/taggers:
    get:
      tags: [Taggers]
      operationId: listTaggers
      summary: List the available taggers
      description: >-
        The versions the tagging engine serves, as it reports them at its `/version`: one for a plain tagger, one
        per tagger when the engine is a router over several taggers. A collection is tagged with one of them
        (`tagger_version` of *Create a collection*). Takes no parameters.
      responses:
        "200":
          description: The tagger versions.
          content:
            application/json:
              schema:
                type: object
                required: [taggers]
                properties:
                  taggers:
                    type: array
                    items: { type: string }
              example:
                taggers: [grep, "decisions/openai:zai-org/GLM-5.3-Flash"]
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "502":
          description: "`tag_engine_error`: the tagging engine's versions could not be fetched."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/collections:
    get:
      tags: [Collections]
      operationId: listCollections
      summary: List collections
      description: >-
        Collections are listed newest first, at most `limit` per page (default 100, at most 1000). When more
        follow, the answer carries `next`: pass it as `cursor` to get the next page, and repeat until an
        answer has no `next`.
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200":
          description: A page of collections. `collections` is `null` when there are none.
          content:
            application/json:
              schema:
                type: object
                properties:
                  collections:
                    type: array
                    nullable: true
                    items:
                      $ref: "#/components/schemas/Collection"
                  next:
                    type: string
                    description: Present only when more collections follow. Pass it as `cursor`.
              example:
                collections:
                  - { id: 5b0c5bd0-5c4e-4a0b-9d4a-6f6a0f3d9a11, name: photos, tagger_version: "decisions/openai:zai-org/GLM-5.3-Flash", created_at: "2026-06-13T12:05:00Z" }
                  - { id: 0a5e1d0e-3b1e-4f43-8f0c-0c1d2e3f4a5b, name: jobs, tagger_version: grep, created_at: "2026-06-13T12:00:00Z" }
                next: azE3ODEzMDYzMDAwMDAwMHw1YjBjNWJkMC01YzRlLTRhMGItOWQ0YS02ZjZhMGYzZDlhMTE
        "400":
          description: "`invalid_parameter`, `unexpected_body`, `invalid_limit` or `invalid_cursor`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/NotReady" }
    post:
      tags: [Collections]
      operationId: createCollection
      summary: Create a collection
      description: |
        A collection is tagged with one tagger version (see *Tagger versions* above); *List the available
        taggers* shows the choices. `tagger_version` may be left out when the tagging engine serves exactly one
        version, which the collection then gets; when it serves several, leaving it out is
        `400 tagger_version_required` and `details.available` lists them. A version that is not served can be
        given too: such a collection can be queried with `evaluate=false`, and evaluated once a tagger of that
        version runs. Objects are plain bytes of any kind, so a collection has no data type.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "400":
          description: "`invalid_json`, `invalid_parameter`, `invalid_name`, `invalid_tagger_version` or `tagger_version_required` (several taggers are served: `details.available` lists them)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: "`already_exists`: a collection with this name exists."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: "`tag_engine_error`: `tagger_version` was left out and the tagging engine's versions could not be fetched."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/collections/{collection}:
    parameters:
      - $ref: "#/components/parameters/Collection"
    delete:
      tags: [Collections]
      operationId: deleteCollection
      summary: Delete a collection
      description: Deletes the collection together with all its objects, tags and stored payloads.
      responses:
        "204":
          description: Deleted
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/collections/{collection}/tags:
    parameters:
      - $ref: "#/components/parameters/Collection"
    get:
      tags: [Tags]
      operationId: listCollectionTags
      summary: Object count and registered tags of a collection
      description: |
        Returns the number of objects in the collection and every tag registered in it, with how
        many objects the tag is known true for, known false for, and not yet evaluated for.
        Tags are ordered by name (byte order). A tag stays registered once seen, even if all objects
        carrying it are later deleted. Expired objects are excluded even before they are removed.
        Tags are listed at most `limit` per page (default 100, at most 1000); when more follow, the answer
        carries `next`: pass it as `cursor` to get the next page. `prefix` keeps only tags that start with it.
      parameters:
        - name: prefix
          in: query
          description: Only tags starting with this literal prefix (at most 128 bytes).
          schema: { type: string, maxLength: 128 }
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200":
          description: Counts and tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionTags"
        "400":
          description: "`invalid_collection_name`, `invalid_parameter`, `unexpected_body`, `invalid_limit`, `invalid_prefix` or `invalid_cursor`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/collections/{collection}/objects/query:
    parameters:
      - $ref: "#/components/parameters/Collection"
    post:
      tags: [Objects]
      operationId: queryObjects
      summary: Query objects by tags
      description: |
        Returns objects ordered by `date` (newest first), then `id`. Tag filters are ANDed. Tags that
        are not known for a candidate object are evaluated by the tagging engine and stored, unless
        `evaluate` is `false`.

        With `evaluate: false` the tagging engine is never called and the query is answered from
        known tags only: an object is returned only if every requested tag is known for it and
        matches. Objects whose requested tags were never evaluated are left out, so the result is a
        subset of what an evaluating query returns. No new tags are registered.

        **Evaluation needs the right tagger.** Evaluating asks the tagging engine for the collection's
        `tagger_version`; if the engine runs another one, the query fails with `409 tagger_version_mismatch`
        (see *Tagger versions* above). `evaluate: false` still answers from the tags the collection has.

        **Page size and cursor.** The answer holds at most `limit` objects (default `TAGONA_DEFAULT_LIMIT`, 5; at most
        `TAGONA_MAX_LIMIT`, 100). When more objects match, it carries `next`: send the same query again
        with `cursor` set to it to get the following page, and repeat until an answer has no `next`. Keep
        `tags`, `date` and `evaluate` the same on every page. The cursor marks a position in the order
        (`date`, then `id`), so objects uploaded or deleted between pages do not shift it. With
        `best_effort`, a query that times out also answers with a `next` to continue from.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagsQueryRequest"
            examples:
              firstPage:
                summary: First page
                value:
                  tags: { golang: true, qa: false }
                  limit: 5
                  evaluate: true
              nextPage:
                summary: The next page, with the `next` of the previous answer
                value:
                  tags: { golang: true, qa: false }
                  limit: 5
                  evaluate: true
                  cursor: MTc4MTMwNjMwMDAwMHwzYWIxYjZlMC1hMTc5LTQyNGItOTgyNS0yZDI4MTdjNzYxYWU
      responses:
        "200":
          description: Matching objects and, when more exist, a cursor for the next page.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TagsQueryResponse" }
              example:
                objects:
                  - id: 3ab1b6e0-a179-424b-9825-2d2817c761ae
                    collection: jobs
                    date: "2026-06-07T12:00:00Z"
                    size_bytes: 52
                    content_hash: 9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08
                    created_at: "2026-06-07T12:00:01Z"
                    metadata: { name: senior-go-engineer.txt }
                    tags: { golang: true, qa: false }
                next: MTc4MTMwNjMwMDAwMHwzYWIxYjZlMC1hMTc5LTQyNGItOTgyNS0yZDI4MTdjNzYxYWU
        "400":
          description: >-
            `invalid_json`, `invalid_parameter`, `invalid_limit`, `invalid_cursor`, `invalid_tags`,
            `invalid_date_filter`, `invalid_timeout` or `invalid_collection_name`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/TaggerVersionMismatch" }
        "500":
          description: "`query_timeout`: the query did not finish within `timeout_ms` (and `best_effort` was off)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: "`tag_engine_error`: the tagging engine failed."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/collections/{collection}/objects/{id}:
    parameters:
      - $ref: "#/components/parameters/Collection"
      - $ref: "#/components/parameters/ObjectId"
    get:
      tags: [Objects]
      operationId: getObject
      summary: Get an object
      description: The object's properties, including its metadata (not its payload, which is `…/data`).
      responses:
        "200":
          description: The object
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Object" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      tags: [Objects]
      operationId: deleteObject
      summary: Delete an object
      description: Hard delete of the object's record, metadata, tags and stored payload.
      responses:
        "204":
          description: Deleted
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/collections/{collection}/objects/{id}/data:
    parameters:
      - $ref: "#/components/parameters/Collection"
      - $ref: "#/components/parameters/ObjectId"
    get:
      tags: [Objects]
      operationId: getObjectData
      summary: Download an object's payload
      description: >-
        Redirects (`307`) to a short-lived, signed URL of the object store, which serves the raw payload
        (`application/octet-stream`) and supports `Range` requests; the payload does not pass through Tagona.
        Follow the redirect without sending the API key on to the object store: the URL itself is the
        credential, valid for `TAGONA_DATA_URL_TTL` (60 seconds by default, and never beyond the object's
        expiry), and for this one payload and the GET method only. Do not cache or share it; ask again when
        it has expired. If the object's metadata has `content_type` (a media type) the store answers with that
        `Content-Type`, and if it has `name` the download is an `attachment` with that file name (cleaned: no
        directories or control characters); a type without a name is a plain `attachment`, so it is never rendered
        from the store's origin. Values that are not usable are ignored, leaving `application/octet-stream`. The object store must be reachable by the client and allow its origin (CORS) for
        downloads from a browser.
      responses:
        "307":
          description: Where to download the payload from.
          headers:
            Location:
              description: The signed URL.
              schema: { type: string, format: uri }
            Cache-Control:
              schema: { type: string, enum: [no-store] }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "An API key created with `POST /v1/admin/api-keys`, sent as `Authorization: Bearer <key>`."
    adminBasicAuth:
      type: http
      scheme: basic
      description: Admin credentials configured on the server. An API key is rejected on admin endpoints.
  parameters:
    Collection:
      name: collection
      in: path
      required: true
      description: Collection name (1-64 chars, lowercase letters, digits, `_` and `-`, starting with a letter).
      schema:
        type: string
        pattern: "^[a-z][a-z0-9_-]{0,63}$"
        example: jobs
    ObjectId:
      name: id
      in: path
      required: true
      description: The object's `id`.
      schema: { type: string, format: uuid }
    PageLimit:
      name: limit
      in: query
      description: Maximum items per page.
      schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
    PageCursor:
      name: cursor
      in: query
      description: The `next` value of the previous page; leave it out for the first page.
      schema: { type: string }
  responses:
    InvalidRequest:
      description: "`invalid_collection_name`, `invalid_parameter` or `unexpected_body`."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: "`not_found`: the collection, object or key does not exist (expired objects count as missing)."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: >-
        `missing_api_key` (no or malformed Bearer header), `invalid_api_key` (unknown key) or `expired_api_key`
        (the key existed but has expired; get a new one).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotReady:
      description: "`not_ready`: the storage service is unavailable."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: not_found }
            message: { type: string, example: not found }
            details: { type: object, additionalProperties: true }
    CollectionCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, pattern: "^[a-z][a-z0-9_-]{0,63}$", example: jobs }
        tagger_version:
          type: string
          minLength: 1
          maxLength: 128
          description: >-
            The tagger version the collection is tagged with, `<implementation>` or
            `<implementation>:<model>`. May be left out when the tagging engine serves only one version.
          example: "decisions/openai:zai-org/GLM-5.3-Flash"
    Collection:
      type: object
      required: [id, name, tagger_version, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        tagger_version:
          type: string
          description: The version of the tagger the collection is tagged with, `<implementation>` or `<implementation>:<model>`.
        created_at: { type: string, format: date-time }
...
    TagsQueryResponse:
      type: object
      required: [objects]
      properties:
        objects:
          type: array
          items: { $ref: "#/components/schemas/Object" }
        next:
          type: string
          description: >-
            Cursor for the next page: send the query again with this as `cursor`. Absent on the last page.
...
    TagStat:
      type: object
      required: [tag, true_count, false_count, first_seen_at]
      properties:
        tag: { type: string }
        true_count: { type: integer, format: int64, description: Objects the tag is known true for. }
        false_count: { type: integer, format: int64, description: Objects the tag is known false for. }
        first_seen_at: { type: string, format: date-time }
    CollectionTags:
      type: object
      required: [collection, total_objects, tags]
      properties:
        collection: { type: string }
        total_objects:
          type: integer
          format: int64
          description: Objects in the whole collection, not only those matching `prefix`.
        tags:
          type: array
          items: { $ref: "#/components/schemas/TagStat" }
        next:
          type: string
          description: Cursor for the next page; absent on the last page.
...